requirements.txt
asgiref-3.4.1
django-3.2.7
pytz-2021.1
sqlparse-0.4.2
djangorestframework-3.12.4
